Features: Added admin route and change some pattern for better UI/UX and fixed some issuse

+++ app/api/profile/route.js
@@ -143,6 +143,15 @@ export async function DELETE(req) {
       );
     }
 
+    const profile = await Profile.findOne({ _id: id });
+
+    if (!user._id.equals(profile.userId)) {
+      return NextResponse.json(
+        { error: "دسترسی شما به این آگهی محدود شده است!" },
+        { status: 403 }
+      );
+    }
+
     await Profile.deleteOne({ _id: id });
 
     return NextResponse.json({ message: "آگهی با موفقیت حذف شد!" });
